totemip: Add support for local ipv6
Unlike the b4a8872 and 934c47e we don't store
the sin6_scope_ip in the structure totem_ip_address
but rather get it in totemip_totemip_to_sockaddr_convert by the use of totemip_getif_scopeid.
To prevent from calling the getifaddrs in the totemip_getif_scopeid repeatedly,
we use the static struct ifaddrs *totemip_getif_scopeid_ifap. When the corosync-cfgtool -R is called,
the totemip_getif_scopeid will refresh.
If there are several interfaces with the same IP address, it prints the warning in the journal.